Api endpoints skeletron
