+++ app_api/controllers/locations.js
@@ -1,11 +1,33 @@
+const mongoose = require('mongoose');
+const Parchi = mongoose.model('Location');
 
 //PlaceHolder functions to build the app w/o errors
 
-const locationsCreate = (req, res) => { };
-const locationsReadOne = (req, res) => { };
-const locationsUpdateOne = (req, res) => { };
-const locationsListByDistance = (req, res) => { };
-const locationsDeleteOne = (req, res) => { };
+const locationsCreate = (req, res) => {
+    res
+        .status(200)
+        .json({"status":"success"})
+ };
+const locationsReadOne = (req, res) => {
+    res
+        .status(200)
+        .json({"status":"success"})
+ };
+const locationsUpdateOne = (req, res) => { 
+    res
+        .status(200)
+        .json({"status":"success"})
+};
+const locationsListByDistance = (req, res) => {
+    res
+        .status(200)
+        .json({"status":"success"})
+ };
+const locationsDeleteOne = (req, res) => {
+    res
+        .status(200)
+        .json({"status":"success"})
+ };
  
 
 module.exports= {
