+++ app_api/controllers/reviews.js
@@ -1,10 +1,22 @@
 
 //PlaceHolder functions to build the app w/o errors
 
-const reviewsCreate = (req, res) => { };
-const reviewsReadOne = (req, res) => { };
-const reviewsUpdateOne = (req, res) => { };
-const reviewsDeleteOne = (req, res) => { };
+const reviewsCreate = (req, res) => {
+    res
+        .status(200)
+        .json({"status":"success"})
+ };
+const reviewsReadOne = (req, res) => {res
+    .status(200)
+    .json({"status":"success"}) };
+const reviewsUpdateOne = (req, res) => {res
+    .status(200)
+    .json({"status":"success"}) };
+const reviewsDeleteOne = (req, res) => {
+    res
+        .status(200)
+        .json({"status":"success"})
+ };
  
 
 module.exports= {
